moves sine_dewarp to modules, does not have schema
