+++ tests/modules/sine_dewarp/resources/dewarping_regression_test_output.h5
